fix(bridge): X509_NAME_add_entry_by_txt len is ossl_ssize_t — Win64 stack-slot width mismatch
C++ (x509.h:1447) takes ossl_ssize_t (=ptrdiff_t, 64-bit) for len; the
hand-written Rust extern declared c_int. On Win64 the 5th argument is
passed in a STACK slot and the 32-bit store leaves the upper half as
stack garbage — the callee read garbage<<32|12 as ossl_ssize_t,
CBS_init scanned past the CN literal unbounded, nchar blew past
maxsize=64 → STRING_TOO_LONG. Manifestation depended on prior stack use
(cold vs warm call paths), so every Windows fixture cert intermittently
failed: tls_sni ×6 / proxy_tunnel / ws_tls ×2 / wss (the 'TLS cold-start
flake family' — not RNG at all). Linux SysV passes arg 5 in r8
(zero-extending register write) — always clean, hence Windows-only.

Width audit of the remaining hand-written extern blocks in the tree
(bridge lib/client/session_cache/connection + src/boringssl extra
surface): clean — c_long against long (32-bit under ms_abi), pointers
and ints aligned; this was the sole ossl_ssize_t site.

Evidence (.200 real machine, fixed binary): the cert-build flake is
gone — tls_sni static_cert_regression 4/4 stable (was fail→pass→pass
flaky); Linux bao_boringssl_bridge+bun_boringssl 15/15.

Remaining tls-family failures are a DISTINCT root (client 'data' echo
never delivers while handshake/cert/write all pass) — tracked
separately.

+++ src/bao_boringssl_bridge/src/lib.rs
@@ -130,7 +130,17 @@ unsafe extern "C" {
         field: *const c_char,
         ty: c_int,
         bytes: *const u8,
-        len: c_int,
+        // `ossl_ssize_t` (= ptrdiff_t, 64-bit) in x509.h:1447. Declared
+        // `c_int` until 2026-09-25: on Win64 the 5th argument lives in a
+        // STACK slot and the 32-bit store leaves the upper half as stack
+        // garbage — the callee read `garbage<<32 | 12` as ossl_ssize_t,
+        // CBS_init scanned past the CN literal unbounded, nchar blew past
+        // maxsize=64 → STRING_TOO_LONG. Flaky per call history (the garbage
+        // depends on prior stack use): every Windows fixture cert
+        // intermittently failed (tls_sni ×6 / proxy_tunnel / ws_tls ×2 /
+        // wss). Linux SysV passes arg 5 in r8 (zero-extending register
+        // write) — always clean, which is why the bug was Windows-only.
+        len: isize,
         loc: c_int,
         set: c_int,
     ) -> c_int;
@@ -253,7 +263,7 @@ pub fn generate_self_signed_pem(cn: &str, days: c_long) -> Result<(String, Strin
             c"CN".as_ptr(),
             MBSTRING_ASC,
             cn.as_ptr(),
-            cn.len() as c_int,
+            cn.len() as isize,
             -1,
             0,
         ) == 1
